+++ src/main/java/com/grepp/moodlink/app/model/admin/music/repository/AdminMusicGenreRepository.java
@@ -1,8 +1,7 @@
-package com.grepp.moodlink.app.model.admin.music;
+package com.grepp.moodlink.app.model.admin.music.repository;
 
 import com.grepp.moodlink.app.model.data.music.entity.MusicGenre;
 import java.util.List;
-import java.util.zip.ZipFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminMusicGenreRepository extends JpaRepository<MusicGenre, Long> {
