+++ src/main/java/com/grepp/moodlink/app/model/admin/music/repository/AdminMusicRepository.java
@@ -1,11 +1,7 @@
-package com.grepp.moodlink.app.model.admin.music;
+package com.grepp.moodlink.app.model.admin.music.repository;
 
 import com.grepp.moodlink.app.model.data.music.entity.Music;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
