+++ src/main/java/com/grepp/moodlink/app/model/admin/music/repository/AdminMusicRepositoryCustom.java
@@ -1,6 +1,5 @@
-package com.grepp.moodlink.app.model.admin.music;
+package com.grepp.moodlink.app.model.admin.music.repository;
 
-import com.grepp.moodlink.app.model.data.music.dto.MusicDto;
 import com.grepp.moodlink.app.model.data.music.entity.Music;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
